Move bulky install code to its own file
